Option --dump-every-bb now takes a 64bit value on user request. Update/clarify documentation.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@5806

+++ callgrind/clo.c
@@ -618,7 +618,7 @@ Bool CLG_(process_cmd_line_option)(Char* arg)
        CLG_(clo).dump_bb = False;
 
    else if (0 == VG_(strncmp)(arg, "--dump-every-bb=", 16))
-        CLG_(clo).dump_every_bb = (Int)VG_(atoll)(&arg[16]);
+        CLG_(clo).dump_every_bb = (ULong)VG_(atoll)(&arg[16]);
 
 
    else if (0 == VG_(strcmp)(arg, "--collect-alloc=yes"))
